+++ RELEASE-NOTES.md
@@ -22,7 +22,7 @@
 * Added `--single-transaction` flag to `snow sql`. Wraps BEGIN/COMMIT around statements to execute them as a single transaction, ensuring all commands complete successfully or no change is applied.
 
 ## Fixes and improvements
-
+* Move `enable_release_channels` from global feature flag to project level.
 
 # v3.8.0